+++ src/devices/WN67.ts
@@ -12,8 +12,8 @@ import * as utils from './../Utils';
 
 export class WN67 extends EcowittAccessory {
   static readonly properties: string[] = ['temperature', 'humidity', 'windDirection',
-    'windSpeed', 'windGustSpeed', 'windMaxDailySpeed', 'rainEventTotal', 'rainHourlyTotal',
-    'rainDailyTotal', 'rainWeekyTotal', 'rainMonthlyTotal', 'rainYearlyTotal'];
+    'windSpeed', 'windGustSpeed', 'windMaxDailySpeed', 'rainRate', 'rainEventTotal',
+    'rainHourlyTotal', 'rainDailyTotal', 'rainWeekyTotal', 'rainMonthlyTotal', 'rainYearlyTotal'];
 
   protected battery: BatterySensor | undefined;
   protected temperature: TemperatureSensor | undefined;
@@ -29,6 +29,7 @@ export class WN67 extends EcowittAccessory {
   protected weeklyRain: RainSensor | undefined;
   protected monthlyRain: RainSensor | undefined;
   protected yearlyRain: RainSensor | undefined;
+  protected totalRain: RainSensor | undefined;
 
   constructor(
     protected readonly platform: EcowittPlatform,
@@ -40,8 +41,7 @@ export class WN67 extends EcowittAccessory {
       'wh65batt', 'tempf', 'humidity', 'winddir', 'windspeedmph', 'windgustmph',
       'maxdailygust', 'eventrainin', 'hourlyrainin', 'dailyrainin', 'weeklyrainin',
       'monthlyrainin'];
-    this.optionalData = ['yearlyrainin', 'rainratein'];
-    this.unusedData = ['totalrainin'];
+    this.optionalData = ['yearlyrainin', 'rainratein', 'totalrainin'];
 
     const hideConfig = this.platform.config?.hidden || {};
     const hideConfigCustom = this.platform.config?.customHidden || [];
@@ -175,6 +175,15 @@ export class WN67 extends EcowittAccessory {
       this.yearlyRain.removeService();
       this.yearlyRain = undefined;
     }
+
+    if (!utils.includesAny(hidden, ['rainTotal', `${this.shortServiceId}:rainTotal`])) {
+      nameOverride = utils.lookup(this.platform.config?.nameOverrides, `${this.shortServiceId}:rainTotal`);
+      this.totalRain = new RainSensor(platform, accessory, `${this.accessoryId}:rainTotal`, nameOverride || 'Rain Total');
+    } else {
+      this.totalRain = new RainSensor(platform, accessory, `${this.accessoryId}:rainTotal`, 'Rain Total');
+      this.totalRain.removeService();
+      this.totalRain = undefined;
+    }
   }
 
   //----------------------------------------------------------------------------
@@ -224,6 +233,7 @@ export class WN67 extends EcowittAccessory {
       dataReport.dateutc,
     );
 
+    // optional
     if (dataReport.rainratein === undefined) {
       this.rainRate?.removeService();
       this.rainRate = undefined;
@@ -265,6 +275,7 @@ export class WN67 extends EcowittAccessory {
       dataReport.dateutc,
     );
 
+    // optional
     if (dataReport.yearlyrainin === undefined) {
       this.yearlyRain?.removeService();
       this.yearlyRain = undefined;
@@ -272,6 +283,18 @@ export class WN67 extends EcowittAccessory {
       this.yearlyRain?.updateTotal(
         parseFloat(dataReport.yearlyrainin),
         utils.lookup(this.platform.config?.thresholds, 'rainYearlyTotal'),
+        dataReport.dateutc,
+      );
+    }
+
+    // optional
+    if (dataReport.totalrainin === undefined) {
+      this.totalRain?.removeService();
+      this.totalRain = undefined;
+    } else {
+      this.totalRain?.updateTotal(
+        parseFloat(dataReport.totalrainin),
+        utils.lookup(this.platform.config?.thresholds, 'rainTotal'),
         dataReport.dateutc,
       );
     }