+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 * Fix Plugin UI validation issue
 * Update dependencies and node v22 and node v24 compatibility support
+* Introduce support for Ecowitt Laser Distance Sensor (LDS01)
+* Display service for 'totalRain' data property if present in data report
 
 ### 2.10.2
 
@@ -12,6 +14,8 @@
 * Resolve several issues with error handling and data report parsing
 
 ### 2.10.0
+
+s
 
 * Add support for most Ambient Weather stations and sensors
 * Add logic to create accessory for base station even if station model is not recognized